move timeout guard to sessions
